28-07

+++ admin/modules/index/models/debug.php
@@ -27,24 +27,26 @@ class Model extends \Kotchasan\KBase
   {
     // referer, session, admin
     if (self::$request->initSession() && self::$request->isReferer() && $login = Login::isAdmin()) {
-      // action
-      $action = self::$request->post('action')->toString();
-      // file debug
-      $debug = ROOT_PATH.DATA_FOLDER.'logs/error_log.php';
-      if (is_file($debug)) {
-        if ($action == 'get') {
-          // อ่าน debug
-          $t = self::$request->post('t')->toString();
-          foreach (file($debug) as $i => $row) {
-            if (preg_match('/^\[([0-9\-:\s]+)\][\s]+([A-Z]+):[\s]+(.*)/', trim($row), $match)) {
-              if ($match[1] > $t) {
-                echo "$match[1]\t$match[2]\t$match[3]\n";
+      if ($login['email'] != 'demo' && empty($login['fb'])) {
+        // action
+        $action = self::$request->post('action')->toString();
+        // file debug
+        $debug = ROOT_PATH.DATA_FOLDER.'logs/error_log.php';
+        if (is_file($debug)) {
+          if ($action == 'get') {
+            // อ่าน debug
+            $t = self::$request->post('t')->toString();
+            foreach (file($debug) as $i => $row) {
+              if (preg_match('/^\[([0-9\-:\s]+)\][\s]+([A-Z]+):[\s]+(.*)/', trim($row), $match)) {
+                if ($match[1] > $t) {
+                  echo "$match[1]\t$match[2]\t$match[3]\n";
+                }
               }
             }
+          } elseif ($action == 'clear') {
+            // ลบไฟล์ debug
+            unlink($debug);
           }
-        } elseif ($action == 'clear') {
-          // ลบไฟล์ debug
-          unlink($debug);
         }
       }
     }
